Refactor device errors and pso specialization
