+++ src/backend/dx12/src/lib.rs
@@ -489,7 +489,7 @@ impl q::CommandQueue<Backend> for CommandQueue {
         Ok(None)
     }
 
-    fn wait_idle(&self) -> Result<(), error::HostExecutionError> {
+    fn wait_idle(&self) -> Result<(), error::OutOfMemory> {
         self.raw.signal(self.idle_fence, 1);
         assert_eq!(
             winerror::S_OK,
