+++ src/backend/gl/src/queue.rs
@@ -1170,7 +1170,7 @@ impl hal::queue::CommandQueue<Backend> for CommandQueue {
         Ok(None)
     }
 
-    fn wait_idle(&self) -> Result<(), error::HostExecutionError> {
+    fn wait_idle(&self) -> Result<(), error::OutOfMemory> {
         unsafe {
             self.share.context.finish();
         }
