+++ src/backend/metal/src/command.rs
@@ -2310,7 +2310,7 @@ impl hal::queue::CommandQueue<Backend> for CommandQueue {
         Ok(None)
     }
 
-    fn wait_idle(&self) -> Result<(), error::HostExecutionError> {
+    fn wait_idle(&self) -> Result<(), error::OutOfMemory> {
         QueueInner::wait_idle(&self.shared.queue);
         Ok(())
     }
