+++ src/hal/src/pso/compute.rs
@@ -1,7 +1,9 @@
 //! Compute pipeline descriptor.
 
-use super::{BasePipeline, EntryPoint, PipelineCreationFlags};
-use crate::Backend;
+use crate::{
+    pso::{BasePipeline, EntryPoint, PipelineCreationFlags},
+    Backend,
+};
 
 /// A description of the data needed to construct a compute pipeline.
 #[derive(Debug)]
