+++ src/hal/src/pso/output_merger.rs
@@ -2,8 +2,10 @@
 //! The final stage in a pipeline that creates pixel colors from
 //! the input shader results, depth/stencil information, etc.
 
-use super::graphics::StencilValue;
-use super::State;
+use crate::pso::{
+    graphics::StencilValue,
+    State,
+};
 
 /// A pixel-wise comparison function.
 #[derive(Clone, Copy, Debug, Eq, Hash, Ord, PartialEq, PartialOrd)]
